\name{Icens-internal}
\alias{BVmacprod}
\alias{EMICMmac}
\alias{Intersection}
\alias{Subset}
\alias{VEMICMmac}   
\alias{rescaleP}

\title{Internal Icens functions}
\description{
 Internal Icens functions
}
\usage{
These are not really user level functions.
}
\details{
  These are not to be called by the user.
}
\keyword{ts}
