\name{pruitt}
\alias{pruitt}
\non_function{}
\title{ A small artificial, bivariate right-censored data set. }
\description{
The \code{pruitt} data was given in Pruitt (1993) as an example for
testing different methods of estimating the bivariate NPMLE for right
censored data.
This matrix represents the clique matrix of the intersection graph of
the data set given by Pruitt.
}
\format{
This data frame contains 
8 columns, labeled A through H that represent the observations. There
are seven rows corresponding to the seven maximal cliques in the
intersection graph.
}
\source{
Small Sample Comparison of Six Bivariate Survival Curve Estimators,
Journal of Statistical Computation and Simulation, R. Pruitt, 1993.
}
\examples{
data(pruitt)
}
\keyword{datasets}
